Introduce alpha version of patch generator/applier
 1. the generator consists of a simple Shell script
 2. the applier is written in Java and executes commands in batch

The whole idea is to apply patches to local installations while
remaining simple (thus excluding binary diff approach e.g.).

The generator does not need to be portable as it is executed from
a single place (therefore: bash). This will produce a batch file,
whose syntax matches the one from [JarUp](https://github.com/xhanin/jarup)
project.

The applier consumes the aforementioned file against a specified
directory.

This is of course an alpha version as there are some improvements
to be made such as:

 - make the apply operation more atomic
 - make the generator more configurable and user-friendly